// generation/WeaponStats/generated/mapping.ts
// This file is generated by the generateWeaponStat.ts script.
// Do not manually make changes to this file.

import { WeaponStat } from "@d2e/types/WeaponStat";

export const WeaponStatIdToWeaponStatMapping: Record<number, WeaponStat> = {
  [155624089]: {
    hash: 155624089,
    name: "Stability",
  },
  [209426660]: {
    hash: 209426660,
    name: "Guard Resistance",
  },
  [360359141]: {
    hash: 360359141,
    name: "Durability",
  },
  [447667954]: {
    hash: 447667954,
    name: "Draw Time",
  },
  [925767036]: {
    hash: 925767036,
    name: "Ammo Capacity",
  },
  [943549884]: {
    hash: 943549884,
    name: "Handling",
  },
  [1240592695]: {
    hash: 1240592695,
    name: "Range",
  },
  [1345609583]: {
    hash: 1345609583,
    name: "Aim Assistance",
  },
  [1546607977]: {
    hash: 1546607977,
    name: "Heroic Resistance",
  },
  [1591432999]: {
    hash: 1591432999,
    name: "Accuracy",
  },
  [1842278586]: {
    hash: 1842278586,
    name: "Shield Duration",
  },
  [1931675084]: {
    hash: 1931675084,
    name: "Inventory Size",
  },
  [2523465841]: {
    hash: 2523465841,
    name: "Velocity",
  },
  [2714457168]: {
    hash: 2714457168,
    name: "Airborne Effectiveness",
  },
  [2715839340]: {
    hash: 2715839340,
    name: "Recoil Direction",
  },
  [2762071195]: {
    hash: 2762071195,
    name: "Guard Efficiency",
  },
  [2837207746]: {
    hash: 2837207746,
    name: "Swing Speed",
  },
  [2961396640]: {
    hash: 2961396640,
    name: "Charge Time",
  },
  [3017642079]: {
    hash: 3017642079,
    name: "Boost",
  },
  [3022301683]: {
    hash: 3022301683,
    name: "Charge Rate",
  },
  [3555269338]: {
    hash: 3555269338,
    name: "Zoom",
  },
  [3597844532]: {
    hash: 3597844532,
    name: "Precision Damage",
  },
  [3614673599]: {
    hash: 3614673599,
    name: "Blast Radius",
  },
  [3736848092]: {
    hash: 3736848092,
    name: "Guard Endurance",
  },
  [3871231066]: {
    hash: 3871231066,
    name: "Magazine",
  },
  [3907551967]: {
    hash: 3907551967,
    name: "Move Speed",
  },
  [3988418950]: {
    hash: 3988418950,
    name: "Time to Aim Down Sights",
  },
  [4043523819]: {
    hash: 4043523819,
    name: "Impact",
  },
  [4188031367]: {
    hash: 4188031367,
    name: "Reload Speed",
  },
  [4284893193]: {
    hash: 4284893193,
    name: "Rounds Per Minute",
  },
};